requirements.txt
pyaudio
openai
websocket-client
python-dotenv
requests